Create RockPaperScissors.kt
This program first defines an array of three options: rock, paper, and scissors. It then generates a random choice for the computer using the random() function. It prompts the user to enter their choice and reads it using readLine(). It then compares the user's choice to the computer's choice using a series of if-else statements and prints the result to the console.

Screen Demo :
Enter your choice (rock, paper, or scissors): rock
Computer chooses paper
Computer wins!