SaveImageToInternalStorage
This code defines a function called saveImageToInternalStorage that takes a Bitmap object as input and returns a Uri object. The function attempts to save the Bitmap to the app's private external storage directory in the Pictures folder with the filename "photo.jpg".

The function first creates a new File object by calling the getExternalFilesDir method on the Activity and passing in the Environment.DIRECTORY_PICTURES constant to specify that the file should be saved in the Pictures folder. The filename is set to "photo.jpg".

Next, the function attempts to save the Bitmap to the file by creating a new FileOutputStream and passing in the File object. The Bitmap is compressed to a JPEG format with a quality of 100 using the compress method, and the resulting stream is flushed and closed.

If an IOException occurs during this process, the function will print the stack trace of the exception.

Finally, the function returns a Uri object that points to the saved file by concatenating the absolute path of the File object and the filename of the Bitmap object separated by a file separator character. However, this is incorrect and can lead to an invalid URI being returned. A correct implementation would return a Uri object that points to the saved file by calling FileProvider.getUriForFile() method with the file and a content provider authority.