Create DownloadFileFromUri.kt
The code provided is written in Kotlin and contains a function named downloadIntroAudio that downloads an audio file from a given URL using the PRDownloader library. The downloaded file is saved in a specified directory with a given filename.

In the main function, the URL of the audio file is set to soundUrl, and the filename for the downloaded audio file is set to filename. The directory where the audio file is saved is specified as dir, which is created as a File object using the requireActivity() method. Finally, the downloadIntroAudio function is called with the given parameters to initiate the download of the audio file.

Overall, this code downloads an audio file from a URL and saves it to a specified directory with a given filename using the PRDownloader library in Kotlin.