LoadingDialog
The showDialog method is used to show a dialog box on top of the current screen. The first parameter context is required and is the context of the current widget. The barrierDismissible parameter is optional and specifies whether the user can dismiss the dialog box by tapping outside of it. If it is set to false, the user cannot dismiss the dialog box until the operation that is being performed is complete.